Update dependencies, Add changelog, Rename test_constants to tags
