+++ examples/todoapp/integration_test/tests/edit_todo_page_test.dart
@@ -1,4 +1,4 @@
-import '../main/base/test_constants.dart';
+import '../main/base/tags.dart';
 import '../main/base/utils.dart';
 import '../main/configuration/build_app.dart';
 import '../main/configuration/patrol_base_config.dart';
