+++ examples/todoapp/integration_test/tests/statistics_page_test.dart
@@ -1,6 +1,6 @@
 import 'package:todoapp/lib_todo_actions/models/bulk_action.dart';
 
-import '../main/base/test_constants.dart';
+import '../main/base/tags.dart';
 import '../main/base/utils.dart';
 import '../main/configuration/build_app.dart';
 import '../main/configuration/patrol_base_config.dart';
