+++ examples/todoapp/integration_test/tests/todo_list_test.dart
@@ -1,7 +1,7 @@
 import 'package:todoapp/base/models/todos_filter_model.dart';
 import 'package:todoapp/lib_todo_actions/models/bulk_action.dart';
 
-import '../main/base/test_constants.dart';
+import '../main/base/tags.dart';
 import '../main/base/utils.dart';
 import '../main/configuration/build_app.dart';
 import '../main/configuration/patrol_base_config.dart';
