+++ packages/rx_bloc_cli/mason_templates/bricks/feature_counter/__brick__/{{#enable_patrol}}integration_test{{/enable_patrol}}/tests/counter_test.dart
@@ -1,6 +1,6 @@
 {{> licence.dart }}
 
-import '../main/base/test_constants.dart';
+import '../main/base/tags.dart';
 import '../main/configuration/build_app.dart';
 import '../main/configuration/patrol_base_config.dart';
 import '../main/steps_utils/counter_page_steps.dart';
