+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/pubspec.yaml
@@ -8,54 +8,53 @@ environment:
   flutter: '>=3.13.0 <4.0.0'
 
 dependencies: {{#enable_dev_menu}}
-  alice: ^1.0.0-dev.10
-  alice_dio: ^1.0.4{{/enable_dev_menu}}{{#enable_patrol}}
-  args: ^2.5.0{{/enable_patrol}}
+  alice: ^1.0.0-dev.12
+  alice_dio: ^1.0.6{{/enable_dev_menu}}
   collection: ^1.18.0
   copy_with_extension: ^5.0.4{{#enable_pin_code}}
-  crypto: ^3.0.3{{/enable_pin_code}}
-  cupertino_icons: ^1.0.6
-  dio: ^5.3.3
+  crypto: ^3.0.6{{/enable_pin_code}}
+  cupertino_icons: ^1.0.8
+  dio: ^5.7.0
   equatable: ^2.0.5{{#analytics}}
-  firebase_analytics: ^11.3.0{{/analytics}}{{#uses_firebase}}
-  firebase_core: ^3.4.0{{/uses_firebase}}{{#analytics}}
-  firebase_crashlytics: ^4.1.0{{/analytics}}{{#push_notifications}}
-  firebase_messaging: ^15.1.0{{/push_notifications}}
+  firebase_analytics: ^11.3.3{{/analytics}}{{#uses_firebase}}
+  firebase_core: ^3.6.0{{/uses_firebase}}{{#analytics}}
+  firebase_crashlytics: ^4.1.3{{/analytics}}{{#push_notifications}}
+  firebase_messaging: ^15.1.3{{/push_notifications}}
   flutter:
     sdk: flutter{{#enable_social_logins}}
-  flutter_facebook_auth: ^7.1.0{{/enable_social_logins}}
-  flutter_local_notifications: ^17.2.1+2
+  flutter_facebook_auth: ^7.1.1{{/enable_social_logins}}
+  flutter_local_notifications: ^17.2.3
   flutter_localizations:
     sdk: flutter
   flutter_rx_bloc: ^7.0.0
-  flutter_secure_storage: ^9.0.0{{#enable_social_logins}}
-  flutter_svg: ^2.0.7{{/enable_social_logins}}
-  go_router: ^14.2.1
+  flutter_secure_storage: ^9.2.2{{#enable_social_logins}}
+  flutter_svg: ^2.0.10{{/enable_social_logins}}
+  go_router: ^14.3.0
   golden_toolkit: ^0.15.0{{#enable_social_logins}}
-  google_sign_in: ^6.1.5{{/enable_social_logins}}
-  http: ^1.1.0 # This package is used for the local server (if not used, can be removed)
+  google_sign_in: ^6.2.1{{/enable_social_logins}}
+  http: ^1.2.2 # This package is used for the local server (if not used, can be removed)
   jaguar_jwt: ^3.0.0 # This package is used for the local server (if not used, can be removed)
-  json_annotation: ^4.8.1
+  json_annotation: ^4.9.0
   jwt_decoder: ^2.0.1{{#enable_pin_code}}
   local_session_timeout: ^3.2.0{{/enable_pin_code}}
-  provider: ^6.0.5
-  retrofit: ^4.0.2
+  provider: ^6.1.2
+  retrofit: ^4.4.1
   rx_bloc: ^6.0.0
   rx_bloc_list: ^5.0.0
   rxdart: ^0.28.0
-  shared_preferences: ^2.2.1
+  shared_preferences: ^2.3.2
   shelf: ^1.4.1 # This package is used for the local server (if not used, can be removed)
   shelf_router: ^1.1.4 # This package is used for the local server (if not used, can be removed)
-  shelf_static: ^1.1.2 # This package is used for the local server (if not used, can be removed){{#enable_social_logins}}
-  sign_in_with_apple: ^6.1.2{{/enable_social_logins}}{{#enable_mfa}}
-  uuid: ^4.2.1{{/enable_mfa}}
+  shelf_static: ^1.1.3 # This package is used for the local server (if not used, can be removed){{#enable_social_logins}}
+  sign_in_with_apple: ^6.1.3{{/enable_social_logins}}{{#enable_mfa}}
+  uuid: ^4.5.1{{/enable_mfa}}
   widget_toolkit: ^0.2.0 {{#enable_pin_code}}
   widget_toolkit_biometrics: ^0.1.0 {{/enable_pin_code}} {{#enable_feature_otp}}
   widget_toolkit_otp: ^0.1.0{{/enable_feature_otp}}{{#enable_pin_code}}
   widget_toolkit_pin: ^0.2.0{{/enable_pin_code}}
 
 dev_dependencies:
-  build_runner: ^2.4.12
+  build_runner: ^2.4.13
   # flutter_driver:
   #   sdk: flutter
   clean_coverage: ^0.0.3
@@ -63,12 +62,12 @@ dev_dependencies:
   flutter_lints: any
   flutter_test:
     sdk: flutter
-  go_router_builder: ^2.7.0
-  json_serializable: ^6.7.1
-  mockito: ^5.4.2 {{#enable_patrol}}
-  patrol: ^3.11.1 {{/enable_patrol}}
+  go_router_builder: ^2.7.1
+  json_serializable: ^6.8.0
+  mockito: ^5.4.4 {{#enable_patrol}}
+  patrol: ^3.11.2 {{/enable_patrol}}
   r_flutter: 0.9.0
-  retrofit_generator: ^9.1.1
+  retrofit_generator: ^9.1.3
   rx_bloc_generator: ^8.0.0
   rx_bloc_test: ^5.0.0
   test: any {{#enable_patrol}}
