+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/pubspec_overrides.yaml
@@ -1,12 +1,1 @@
 dependency_overrides:
-  ## TODO: Temporary fix until [this issue](https://github.com/jhomlala/alice/issues/234) is closed
-  alice:
-    git:
-      url: https://github.com/DDavidPrime/alice.git
-      ref: rxdart-upgrade
-      path: packages/alice
-  alice_dio:
-    git:
-      url: https://github.com/DDavidPrime/alice.git
-      ref: rxdart-upgrade
-      path: packages/alice_dio
